setup_folder/requirements.txt
PySide2
PySide6
pyqt5
GitPython
